+++ emg3d/io.py
@@ -30,7 +30,7 @@ except ImportError:
     h5py = ("'.h5'-files require `h5py`. Install it via\n"
             "`pip install h5py` or `conda install -c conda-forge h5py`.")
 
-from emg3d import fields, maps, models, utils, meshes, surveys, simulations
+from emg3d import fields, models, utils, meshes, surveys, simulations
 
 __all__ = ['save', 'load']
 
@@ -38,7 +38,6 @@ __all__ = ['save', 'load']
 KNOWN_CLASSES = {
     'Model': models.Model,
     'Field': fields.Field,
-    'BaseMap': maps.BaseMap,
     'Survey': surveys.Survey,
     'Dipole': surveys.Dipole,
     'TensorMesh': meshes.TensorMesh,