+++ tests/test_io.py
@@ -3,7 +3,7 @@ import numpy as np
 from copy import deepcopy as dc
 from numpy.testing import assert_allclose
 
-from emg3d import meshes, models, fields, utils, io, surveys, simulations, maps
+from emg3d import meshes, models, fields, utils, io, surveys, simulations
 
 # Soft dependencies
 try:
@@ -197,6 +197,9 @@ def test_compare_dicts(capsys):
     out = io._compare_dicts(dict1, dict2, True)
     assert out is False
     outstr, _ = capsys.readouterr()
+    print(80*'=')
+    print(outstr)
+    print(80*'=')
     assert " True  :: model      > property_x" in outstr
     assert "  {1}  ::              mu_r" in outstr
     assert " False ::              hy" in outstr
@@ -218,12 +221,6 @@ def test_known_classes(tmpdir):
         'Model': model,
         'Field': field,
         'Dipole': pointdip,
-        'MapConductivity': maps.MapConductivity(),
-        'MapLgConductivity': maps.MapLgConductivity(),
-        'MapLnConductivity': maps.MapLnConductivity(),
-        'MapResistivity': maps.MapResistivity(),
-        'MapLgResistivity': maps.MapLgResistivity(),
-        'MapLnResistivity': maps.MapLnResistivity(),
     }
 
     if xarray: